+++ views/index.ejs
@@ -25,13 +25,13 @@
                 </tr>
             </thead>
             <tbody>
-                <% shortUrls.forEach(fshortUrl => { %>
+                <% shortUrls.forEach(shortUrl => { %>
                 <tr>
                     <td><a href="<%= shortUrl.full %>"><%= shortUrl.full %></a></td>
                     <td><a href="<%= shortUrl.short %>"><%= shortUrl.short %></a></td>
                     <td><%= shortUrl.clicks %></td>
                 </tr>
-                <% )} %>
+                <% }) %>
             </tbody>
         </table>
     </main>
